Limit user choices only to the actions they can perform
