+++ opentech/apply/determinations/views.py
@@ -11,44 +11,19 @@ from django.views.generic import DetailView
 from opentech.apply.activity.models import Activity
 from opentech.apply.activity.messaging import messenger, MESSAGES
 from opentech.apply.funds.models import ApplicationSubmission
-from opentech.apply.funds.workflow import DETERMINATION_OUTCOMES
 from opentech.apply.utils.views import CreateOrUpdateView, ViewDispatcher
 from opentech.apply.users.decorators import staff_required
 
 from .forms import ConceptDeterminationForm, ProposalDeterminationForm
-from .models import Determination, TRANSITION_DETERMINATION, DeterminationMessageSettings, NEEDS_MORE_INFO
+from .models import Determination, DeterminationMessageSettings, NEEDS_MORE_INFO
+
+from .utils import can_create_determination, can_edit_determination, transition_from_outcome
 
 
 def get_form_for_stage(submission):
     forms = [ConceptDeterminationForm, ProposalDeterminationForm]
     index = submission.workflow.stages.index(submission.stage)
     return forms[index]
-
-
-def determination_actions(user, submission):
-    return [action[0] for action in submission.get_actions_for_user(user)]
-
-
-def transition_from_outcome(outcome, submission):
-    for transition_name in submission.phase.transitions:
-        try:
-            transition_type = TRANSITION_DETERMINATION[transition_name]
-        except KeyError:
-            pass
-        else:
-            if transition_type == outcome:
-                return transition_name
-
-
-def can_edit_determination(user, determination, submission):
-    outcome = transition_from_outcome(determination.outcome, submission)
-    valid_outcomes = determination_actions(user, submission)
-    return outcome in valid_outcomes
-
-
-def can_create_determination(user, submission):
-    actions = determination_actions(user, submission)
-    return any(action in DETERMINATION_OUTCOMES for action in actions)
 
 
 @method_decorator(staff_required, name='dispatch')
